feat: create otp box size util
